+++ src/main/java/com/uwetrottmann/tmdb2/Tmdb.java
@@ -1,20 +1,3 @@
-/*
- * Copyright 2013 Uwe Trottmann
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * 
- */
-
 package com.uwetrottmann.tmdb2;
 
 import com.uwetrottmann.tmdb2.services.CollectionService;
@@ -28,120 +11,88 @@ import com.uwetrottmann.tmdb2.services.SearchService;
 import com.uwetrottmann.tmdb2.services.TvEpisodesService;
 import com.uwetrottmann.tmdb2.services.TvSeasonsService;
 import com.uwetrottmann.tmdb2.services.TvService;
-import okhttp3.HttpUrl;
 import okhttp3.Interceptor;
 import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.Response;
-import okhttp3.logging.HttpLoggingInterceptor;
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
-
-import java.io.IOException;
 
 /**
  * Helper class for easy usage of the TMDB v3 API using retrofit.
  *
- * <p>Create an instance of this class, {@link #setApiKey(String)} and then call any of the service methods.
+ * <p>Create an instance of this class and then call any of the service methods.
  *
- * <p>The service methods take care of constructing the required {@link Retrofit} instance and creating the service. You
- * can customize by overriding {@link #retrofitBuilder()} and {@link #okHttpClientBuilder()}
+ * <p>The service methods take care of constructing the required {@link Retrofit} instance and creating the service. It
+ * is recommended you use your existing OkHttp client instance by overriding {@link #okHttpClient()}.
  *
  * <p>Only one {@link Retrofit} instance is created upon the first and re-used for any consequent service method call.
  */
-@SuppressWarnings("WeakerAccess")
 public class Tmdb {
 
-    /**
-     * Tmdb API URL.
-     */
-    public static final String API_URL = "https://api.themoviedb.org/3/";
+    public static final String API_HOST = "api.themoviedb.org";
+    public static final String API_VERSION = "3";
+    public static final String API_URL = "https://" + API_HOST + "/" + API_VERSION + "/";
 
     /**
      * API key query parameter name.
      */
     public static final String PARAM_API_KEY = "api_key";
 
+    private OkHttpClient okHttpClient;
     private Retrofit retrofit;
-    private HttpLoggingInterceptor logging;
 
-    private boolean isDebug;
     private String apiKey;
 
     /**
      * Create a new manager instance.
+     *
+     * @param apiKey Your TMDB API key.
      */
-    public Tmdb() {
+    public Tmdb(String apiKey) {
+        this.apiKey = apiKey;
+    }
+
+    public void apiKey(String apiKey) {
+        this.apiKey = apiKey;
+    }
+
+    public String apiKey() {
+        return apiKey;
     }
 
     /**
-     * Set the TMDB API key. <p> The next service method call will trigger a rebuild of the {@link Retrofit} instance.
-     * If you have cached any service instances, get a new one from its service method.
+     * Creates a {@link Retrofit.Builder} that sets the base URL, adds a Gson converter and sets {@link #okHttpClient()}
+     * as its client.
      *
-     * @param value Your TMDB API key.
-     */
-    public Tmdb setApiKey(String value) {
-        this.apiKey = value;
-        retrofit = null;
-        return this;
-    }
-
-    /**
-     * Set the default logging interceptors log level.
-     *
-     * @param isDebug If true, the log level is set to {@link HttpLoggingInterceptor.Level#BODY}. Otherwise {@link
-     * HttpLoggingInterceptor.Level#NONE}.
-     * @see #okHttpClientBuilder()
-     */
-    public Tmdb setIsDebug(boolean isDebug) {
-        this.isDebug = isDebug;
-        if (logging != null) {
-            logging.setLevel(isDebug ? HttpLoggingInterceptor.Level.BODY : HttpLoggingInterceptor.Level.NONE);
-        }
-        return this;
-    }
-
-    /**
-     * Creates a {@link Retrofit.Builder} that sets the base URL, adds a Gson converter and sets {@link
-     * #okHttpClientBuilder()} as its client. <p> Override this to for example set your own call executor.
-     *
-     * @see #okHttpClientBuilder()
+     * @see #okHttpClient()
      */
     protected Retrofit.Builder retrofitBuilder() {
         return new Retrofit.Builder()
                 .baseUrl(API_URL)
                 .addConverterFactory(GsonConverterFactory.create(TmdbHelper.getGsonBuilder().create()))
-                .client(okHttpClientBuilder().build());
+                .client(okHttpClient());
+    }
+
+
+    /**
+     * Returns the default OkHttp client instance. It is strongly recommended to override this and use your app
+     * instance.
+     *
+     * @see #setOkHttpClientDefaults(OkHttpClient.Builder)
+     */
+    protected synchronized OkHttpClient okHttpClient() {
+        if (okHttpClient == null) {
+            OkHttpClient.Builder builder = new OkHttpClient.Builder();
+            setOkHttpClientDefaults(builder);
+            okHttpClient = builder.build();
+        }
+        return okHttpClient;
     }
 
     /**
-     * Creates a {@link OkHttpClient.Builder} for usage with {@link #retrofitBuilder()}. Adds interceptors to add auth
-     * headers and to log requests. <p> Override this to for example add your own interceptors.
-     *
-     * @see #retrofitBuilder()
+     * Adds an interceptor to add the api key query parameter and to log requests.
      */
-    protected OkHttpClient.Builder okHttpClientBuilder() {
-        OkHttpClient.Builder builder = new OkHttpClient.Builder();
-        builder.addInterceptor(new Interceptor() {
-            @Override
-            public Response intercept(Chain chain) throws IOException {
-                Request request = chain.request();
-
-                HttpUrl.Builder urlBuilder = request.url().newBuilder();
-                urlBuilder.addEncodedQueryParameter(PARAM_API_KEY, apiKey);
-
-                Request.Builder builder = request.newBuilder();
-                builder.url(urlBuilder.build());
-
-                return chain.proceed(builder.build());
-            }
-        });
-        if (isDebug) {
-            logging = new HttpLoggingInterceptor();
-            logging.setLevel(HttpLoggingInterceptor.Level.BODY);
-            builder.addInterceptor(logging);
-        }
-        return builder;
+    protected void setOkHttpClientDefaults(OkHttpClient.Builder builder) {
+        builder.addInterceptor(new TmdbInterceptor(this));
     }
 
     /**