+++ src/main/java/com/uwetrottmann/tmdb2/entities/ContentRating.java
@@ -1,7 +1,6 @@
 package com.uwetrottmann.tmdb2.entities;
 
-public class ProductionCountry {
-
+public class ContentRating {
     public String iso_3166_1;
-    public String name;
+    public String rating;
 }