+++ src/main/java/com/uwetrottmann/tmdb2/entities/Movie.java
@@ -1,20 +1,3 @@
-/*
- * Copyright 2012 Uwe Trottmann
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * 
- */
-
 package com.uwetrottmann.tmdb2.entities;
 
 import java.util.Date;
@@ -47,8 +30,9 @@ public class Movie {
     public Integer vote_count;
 
     // Following are used with append_to_response
-    public Videos videos;
-    public ReleaseDatesResults release_dates;
+    public MovieAlternativeTitles alternative_titles;
     public Credits credits;
+    public ReleaseDatesResults release_dates;
     public MovieResultsPage similar;
+    public Videos videos;
 }