+++ src/main/java/com/uwetrottmann/tmdb2/enumerations/AppendToResponseItem.java
@@ -25,7 +25,9 @@ public enum AppendToResponseItem {
     CREDITS("credits"),
     SIMILAR("similar"),
     IMAGES("images"),
-    EXTERNAL_IDS("external_ids");
+    ALTERNATIVE_TITLES("alternative_titles"),
+    EXTERNAL_IDS("external_ids"),
+    CONTENT_RATINGS("content_ratings");
 
     private final String value;