+++ src/test/java/com/uwetrottmann/tmdb2/TestData.java
@@ -1,19 +1,3 @@
-/*
- * Copyright 2014 Uwe Trottmann
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package com.uwetrottmann.tmdb2;
 
 public interface TestData {
@@ -30,7 +14,9 @@ public interface TestData {
 
     String TVSHOW_TITLE = "Breaking Bad";
     int TVSHOW_ID = 1396;
-    int TVSHOW_TVDB_ID = 81189;
+    String TVSHOW_IMDB_ID = "tt0903747";
+    int TVSHOW_SEASON1_ID = 30272;
+    String EPISODE_IDMB_ID = "tt0959621";
     int PERSON_ID = 1;
     String PERSON_NAME = "George Lucas";
     String PERSON_IMDB_ID = "nm0000184";