+++ src/test/java/com/uwetrottmann/tmdb2/services/TvServiceTest.java
@@ -8,6 +8,7 @@ import com.uwetrottmann.tmdb2.entities.ExternalIds;
 import com.uwetrottmann.tmdb2.entities.Images;
 import com.uwetrottmann.tmdb2.entities.Person;
 import com.uwetrottmann.tmdb2.entities.TvAlternativeTitles;
+import com.uwetrottmann.tmdb2.entities.TvContentRatings;
 import com.uwetrottmann.tmdb2.entities.TvKeywords;
 import com.uwetrottmann.tmdb2.entities.TvResultsPage;
 import com.uwetrottmann.tmdb2.entities.TvSeason;
@@ -34,8 +35,13 @@ public class TvServiceTest extends BaseTestCase {
     public void test_tvshow_with_append_to_response() throws IOException {
         Call<TvShowComplete> call = getManager().tvService().tv(
                 TestData.TVSHOW_ID, null,
-                new AppendToResponse(AppendToResponseItem.CREDITS, AppendToResponseItem.EXTERNAL_IDS,
-                        AppendToResponseItem.IMAGES)
+                new AppendToResponse(AppendToResponseItem.CREDITS,
+                        AppendToResponseItem.VIDEOS,
+                        AppendToResponseItem.ALTERNATIVE_TITLES,
+                        AppendToResponseItem.SIMILAR,
+                        AppendToResponseItem.EXTERNAL_IDS,
+                        AppendToResponseItem.IMAGES,
+                        AppendToResponseItem.CONTENT_RATINGS)
         );
         TvShowComplete show = call.execute().body();
         assertTvShow(show);
@@ -57,6 +63,38 @@ public class TvServiceTest extends BaseTestCase {
         assertThat(show.external_ids.tvdb_id).isNotNull();
         assertThat(show.external_ids.imdb_id).isNotNull();
         assertThat(show.external_ids.tvrage_id).isNotNull();
+
+        // similar
+        assertThat(show.similar).isNotNull();
+        assertThat(show.similar.results).isNotNull();
+        assertThat(show.similar.results).isNotEmpty();
+
+        // videos
+        assertThat(show.videos).isNotNull();
+        assertThat(show.videos.results).isNotNull();
+        assertThat(show.videos.results).isNotEmpty();
+
+        // alternative_titles
+        assertThat(show.alternative_titles).isNotNull();
+        assertThat(show.alternative_titles.results).isNotNull();
+        assertThat(show.alternative_titles.results).isNotEmpty();
+
+        // content ratings
+        assertThat(show.content_ratings).isNotNull();
+        assertThat(show.content_ratings.results).isNotEmpty();
+        assertThat(show.content_ratings.results.get(0).iso_3166_1).isNotNull();
+        assertThat(show.content_ratings.results.get(0).rating).isNotNull();
+    }
+
+    @Test
+    public void test_content_ratings() throws IOException {
+        Call<TvContentRatings> call = getManager().tvService().content_ratings(TestData.TVSHOW_ID);
+        TvContentRatings ratings = call.execute().body();
+        assertThat(ratings).isNotNull();
+        assertThat(ratings.id).isEqualTo(TestData.TVSHOW_ID);
+        assertThat(ratings.results).isNotEmpty();
+        assertThat(ratings.results.get(0).iso_3166_1).isNotNull();
+        assertThat(ratings.results.get(0).rating).isNotNull();
     }
 
     @Test
@@ -143,7 +181,7 @@ public class TvServiceTest extends BaseTestCase {
         assertThat(videos.results.get(0).name).isNotNull();
         assertThat(videos.results.get(0).site).isEqualTo("YouTube");
         assertThat(videos.results.get(0).size).isNotNull();
-        assertThat(videos.results.get(0).type).isEqualTo("Opening Credits");
+        assertThat(videos.results.get(0).type).isEqualTo("Trailer");
     }
 
     @Test